des trucs très prometteurs et les polygones

+++ Projet_AL_Alazet_Deguillaume/src/model/CanvasPosition.java
@@ -2,18 +2,18 @@ package model;
 
 public class CanvasPosition implements PositionI {
 
-    private int x, y;
+    private double x, y;
 
-    public CanvasPosition(int x, int y) {
+    public CanvasPosition(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public int getY() {
+    public double getY() {
         return this.y;
     }
 }
